% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createTimeDistanceFunction.R
\name{createTimeDistanceFunction}
\alias{createTimeDistanceFunction}
\title{Creates time_distance function and load sample data}
\usage{
createTimeDistanceFunction(functionSchema = "public", con, loadData = FALSE,
  dataSchema = "public")
}
\arguments{
\item{functionSchema}{Schema where \code{time_distance} function will be created. Default is \code{public}.}

\item{con}{Connection to PostgreSQL database as produced by \code{dbConnect(...)}}

\item{loadData}{When true than tables \code{small_portfolio} and \code{big_portfolio} with sample data will be created.}

\item{dataSchema}{Schema where sample tables should be created. Default is \code{public}.}
}
\description{
Creates time_distance function in given portfolio and optionally loads sample data.
}
\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
pgCon <- dbConnect(drv, user = 'usr', dbname="db", password = 'secret', host = 'localhost', port = 5432)
createTimeDistanceFunction(con = pgCon, loadData = TRUE)
}
}

