% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadSampleData.R
\name{loadSampleData}
\alias{loadSampleData}
\title{Loads sample data}
\usage{
loadSampleData(con, schema = NULL)
}
\arguments{
\item{con}{Connection to PostgreSQL or Oracle database as produced by \code{dbConnect(...)}}

\item{dataSchema}{Schema where sample tables should be created. Default is \code{public} for
PostgreSQL and user schema for Oracle.}
}
\description{
Loads sample data to PostgreSQL or Oracle database
}
\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
pgCon <- dbConnect(drv, user = 'usr', dbname="db", password = 'secret', host = 'localhost', port = 5432)
createTimeDistanceFunction(con = pgCon, loadData = TRUE)
}
}

