\name{PlotVintageData}
\alias{PlotVintageData}
\title{Plot vintage data}
\usage{
  PlotVintageData(data = NULL, x = "distance",
    y = "event_weight_csum_pct", cond = NULL,
    facets = NULL)
}
\arguments{
  \item{data}{}

  \item{x}{}

  \item{y}{}

  \item{cond}{}

  \item{facets}{}
}
\description{
  This function prints hello
}

