% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printVintageData.R
\name{printVintageData}
\alias{printVintageData}
\title{Print Vintage Data}
\usage{
printVintageData(vintageData, columns = c("vintage_unit_weight",
  "vintage_unit_count", "event_weight", "event_weight_pct", "event_weight_csum",
  "event_weight_csum_pct"), result = "print", file = "vintageData.xls",
  digits = getOption("digits"), stacked = FALSE)
}
\arguments{
\item{vintageData}{Result of \code{getVintageData} function.}

\item{columns}{Columns to be printed. Any combination of the following
possible statistics: \code{vintage_unit_weight, vintage_unit_count,
event_weight, event_weight_pct, event_weight_csum, event_weight_csum_pct}.
For every statistic one table will be printed.}

\item{result}{Type of result. By default, data are printed. If \code{xls} is
used than data are written to XLS file.}

\item{file}{Name of XLS file where data should be stored. Makes sense only
when \code{Result} is set to \code{xls}. If no name is given and
\code{Result} is \code{xls} then name \code{VintageData.xls} will be used
as default.}

\item{digits}{Number of significant digits to be used when printing (only
valid for \code{Result='print'})}

\item{stacked}{If \code{TRUE} then all tables will be saved on single sheet
in XLS. Default is \code{FALSE}.}
}
\description{
This function prints results of \code{getVintageData} function. All
\code{Slicers} in results have to be included in plot definition. Data are
pivoted by \code{distance} column.
}
\examples{
\dontrun{
printVintageData(vintageData,"event_weight_csum_pct", result = 'xls')
printVintageData(vintageData,c("event_weight, event_weight_csum_pct"), result = 'xls')
printVintageData(vintageData, result = 'xls', file = 'MyVintageAnalysis')
}
}

