% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smallPortfolio.R
\docType{data}
\name{smallPortfolio}
\alias{smallPortfolio}
\title{Small Portfolio}
\format{A data frame with 10000 rows and 1 variable.}
\description{
This dataset contains insert statements for PostgreSQL database. Not intended for direct manipulation.
}
\details{
\itemize{
  \item smallPortfolio
}
}
\keyword{datasets}

