% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vintageData.R
\docType{data}
\name{vintageData}
\alias{vintageData}
\title{Vintage data}
\format{A sample dataset for demonstration purposes.}
\description{
This dataset contains sample pre-calculated vintage data to play with. This dataset can be directly used in functions
 \code{plotVintageData} and \code{aggregateVintageData}.
}
\details{
\itemize{
  \item Vintage Data
}
}
\keyword{datasets}

