\docType{data}
\name{BigPortfolio}
\alias{BigPortfolio}
\title{Big Portfolio}
\format{A data frame with 10000 rows and 1 variable.}
\description{
  This dataset contains insert statements for PostgreSQL
  database. Not intended for direct manipulation.
}
\details{
  \itemize{ \item BigPortfolio }
}
\keyword{datasets}

