% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotVintageData.R
\name{plotVintageData}
\alias{plotVintageData}
\title{Plot vintage data}
\usage{
plotVintageData(data = NULL, x = "distance", y = "event_weight_csum_pct",
  cond = NULL, facets = NULL)
}
\arguments{
\item{data}{Result of \code{GetVintageData} function.}

\item{x}{Data to be displayed on \code{x} axis. Default is \code{distance}.}

\item{y}{Data to be displayed on \code{y} axis. Default is \code{event_weight_csum_pct}.}

\item{cond}{Variable to be used for conditioning (in-chart group).}

\item{facets}{Formulat to be used for facetting.}
}
\description{
This function plots results of \code{GetVintageData} function. All \code{Slicers} in results have
to be included in plot definition.
}
\examples{
plotVintageData(vintageData, facets='product~region')
plotVintageData(vintageData, cond = 'origination_month', facets='product~region')
plotVintageData(vintageData, cond = 'origination_month', facets='region~product')
plotVintageData(vintageData, x = 'origination_month', cond = 'distance',
                facets='region~product')
}

