\docType{data}
\name{SmallPortfolio}
\alias{SmallPortfolio}
\title{Small Portfolio}
\format{A data frame with 10000 rows and 1 variable.}
\description{
  This dataset contains insert statements for PostgreSQL
  database. Not intended for direct manipulation.
}
\details{
  \itemize{ \item SmallPortfolio }
}
\keyword{datasets}

