\docType{data}
\name{VintageData}
\alias{VintageData}
\title{Vintage data}
\format{A sample dataset for demonstration purposes.}
\description{
  This dataset contains sample pre-calculated vintage data
  to play with. This dataset can be directly used in
  functions \code{PlotVintageData} and
  \code{AggregateVintageData} or used with Shiny
  application \code{pgvintshiny}
}
\details{
  \itemize{ \item Vintage Data }
}
\keyword{datasets}

